// src/common/enums/achievement-type.enum.ts
export enum AchievementType {
  FIRST_SONG = 'first_song',
  PERFECT_SCORE = 'perfect_score',
  STREAK_7 = 'streak_7',
  STREAK_30 = 'streak_30',
  LEVEL_UP = 'level_up',
  SONGS_10 = 'songs_10',
  SONGS_50 = 'songs_50',
  SONGS_100 = 'songs_100',
  PRONUNCIATION_MASTER = 'pronunciation_master',
  PITCH_PERFECT = 'pitch_perfect',
  RHYTHM_KING = 'rhythm_king',
  GENRE_EXPLORER = 'genre_explorer',
}